pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "mistralai>=1.8.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.34.3",
]
